# openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAI API Reference Server",
        "description": "This is an API reference server that mimics openai server.",
        "version": "0.1.0"
    },
    "paths": {
        "/openai/deployments/{deployment_name}/chat/completions": {
            "post": {
                "summary": "Create Chat Completion",
                "operationId": "create_chat_completion_openai_deployments__deployment_name__chat_completions_post",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "deployment_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Deployment Name"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatCompletionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/openai/deployments/{deployment_name}/embeddings": {
            "post": {
                "summary": "Create Embedding",
                "operationId": "create_embedding_openai_deployments__deployment_name__embeddings_post",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "deployment_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Deployment Name"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmbeddingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/openai/deployments/{deployment_name}/completions": {
            "post": {
                "summary": "Create Completion",
                "operationId": "create_completion_openai_deployments__deployment_name__completions_post",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "deployment_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Deployment Name"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompletionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ChatCompletionRequest": {
                "properties": {
                    "messages": {
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/ChatCompletionSystemMessageParam"
                                },
                                {
                                    "$ref": "#/components/schemas/ChatCompletionUserMessageParam"
                                },
                                {
                                    "$ref": "#/components/schemas/ChatCompletionAssistantMessageParam"
                                },
                                {
                                    "$ref": "#/components/schemas/ChatCompletionToolMessageParam"
                                },
                                {
                                    "$ref": "#/components/schemas/ChatCompletionFunctionMessageParam"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Messages"
                    },
                    "model": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model"
                    },
                    "frequency_penalty": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Frequency Penalty",
                        "default": 0.0
                    },
                    "logit_bias": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "type": "number"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Logit Bias"
                    },
                    "logprobs": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Logprobs",
                        "default": false
                    },
                    "top_logprobs": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Top Logprobs",
                        "default": 0
                    },
                    "max_tokens": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Tokens"
                    },
                    "n": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "N",
                        "default": 1
                    },
                    "presence_penalty": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Presence Penalty",
                        "default": 0.0
                    },
                    "response_format": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ResponseFormat"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "seed": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seed"
                    },
                    "stop": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stop"
                    },
                    "stream": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stream",
                        "default": false
                    },
                    "stream_options": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/StreamOptions"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "temperature": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Temperature",
                        "default": 0.7
                    },
                    "top_p": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Top P",
                        "default": 1.0
                    },
                    "tools": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/ChatCompletionToolsParam"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tools"
                    },
                    "tool_choice": {
                        "anyOf": [
                            {
                                "type": "string",
                                "enum": [
                                    "none"
                                ],
                                "const": "none"
                            },
                            {
                                "$ref": "#/components/schemas/ChatCompletionNamedToolChoiceParam"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tool Choice",
                        "default": "none"
                    },
                    "user": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User"
                    },
                    "best_of": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Best Of"
                    },
                    "use_beam_search": {
                        "type": "boolean",
                        "title": "Use Beam Search",
                        "default": false
                    },
                    "top_k": {
                        "type": "integer",
                        "title": "Top K",
                        "default": -1
                    },
                    "min_p": {
                        "type": "number",
                        "title": "Min P",
                        "default": 0.0
                    },
                    "repetition_penalty": {
                        "type": "number",
                        "title": "Repetition Penalty",
                        "default": 1.0
                    },
                    "length_penalty": {
                        "type": "number",
                        "title": "Length Penalty",
                        "default": 1.0
                    },
                    "early_stopping": {
                        "type": "boolean",
                        "title": "Early Stopping",
                        "default": false
                    },
                    "stop_token_ids": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "integer"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stop Token Ids"
                    },
                    "include_stop_str_in_output": {
                        "type": "boolean",
                        "title": "Include Stop Str In Output",
                        "default": false
                    },
                    "ignore_eos": {
                        "type": "boolean",
                        "title": "Ignore Eos",
                        "default": false
                    },
                    "min_tokens": {
                        "type": "integer",
                        "title": "Min Tokens",
                        "default": 0
                    },
                    "skip_special_tokens": {
                        "type": "boolean",
                        "title": "Skip Special Tokens",
                        "default": true
                    },
                    "spaces_between_special_tokens": {
                        "type": "boolean",
                        "title": "Spaces Between Special Tokens",
                        "default": true
                    },
                    "truncate_prompt_tokens": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Truncate Prompt Tokens"
                    },
                    "echo": {
                        "type": "boolean",
                        "title": "Echo",
                        "description": "If true, the new message will be prepended with the last message if they belong to the same role.",
                        "default": false
                    },
                    "add_generation_prompt": {
                        "type": "boolean",
                        "title": "Add Generation Prompt",
                        "description": "If true, the generation prompt will be added to the chat template. This is a parameter used by chat template in tokenizer config of the model.",
                        "default": true
                    },
                    "add_special_tokens": {
                        "type": "boolean",
                        "title": "Add Special Tokens",
                        "description": "If true, special tokens (e.g. BOS) will be added to the prompt on top of what is added by the chat template. For most models, the chat template takes care of adding the special tokens so this should be set to false (as is the default).",
                        "default": false
                    },
                    "documents": {
                        "anyOf": [
                            {
                                "items": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Documents",
                        "description": "A list of dicts representing documents that will be accessible to the model if it is performing RAG (retrieval-augmented generation). If the template does not support RAG, this argument will have no effect. We recommend that each document should be a dict containing \"title\" and \"text\" keys."
                    },
                    "chat_template": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chat Template",
                        "description": "A Jinja template to use for this conversion. If this is not passed, the model's default chat template will be used instead."
                    },
                    "chat_template_kwargs": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chat Template Kwargs",
                        "description": "Additional kwargs to pass to the template renderer. Will be accessible by the chat template."
                    },
                    "guided_json": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "object"
                            },
                            {
                                "$ref": "#/components/schemas/BaseModel"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Guided Json",
                        "description": "If specified, the output will follow the JSON schema."
                    },
                    "guided_regex": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Guided Regex",
                        "description": "If specified, the output will follow the regex pattern."
                    },
                    "guided_choice": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Guided Choice",
                        "description": "If specified, the output will be exactly one of the choices."
                    },
                    "guided_grammar": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Guided Grammar",
                        "description": "If specified, the output will follow the context free grammar."
                    },
                    "guided_decoding_backend": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Guided Decoding Backend",
                        "description": "If specified, will override the default guided decoding backend of the server for this specific request. If set, must be either 'outlines' / 'lm-format-enforcer'"
                    },
                    "guided_whitespace_pattern": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Guided Whitespace Pattern",
                        "description": "If specified, will override the default whitespace pattern for guided json decoding."
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "messages"
                ],
                "title": "ChatCompletionRequest"
            },
            "ChatCompletionSystemMessageParam": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "system"
                        ],
                        "const": "system",
                        "title": "Role"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "content",
                    "role"
                ],
                "title": "ChatCompletionSystemMessageParam"
            },
            "ChatCompletionToolMessageParam": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "tool"
                        ],
                        "const": "tool",
                        "title": "Role"
                    },
                    "tool_call_id": {
                        "type": "string",
                        "title": "Tool Call Id"
                    }
                },
                "type": "object",
                "required": [
                    "content",
                    "role",
                    "tool_call_id"
                ],
                "title": "ChatCompletionToolMessageParam"
            },
...
            "ChatCompletionUserMessageParam": {
                "properties": {
                    "content": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "items": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ChatCompletionContentPartTextParam"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ChatCompletionContentPartImageParam"
                                        }
                                    ]
                                },
                                "type": "array"
                            }
                        ],
                        "title": "Content"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "user"
                        ],
                        "const": "user",
                        "title": "Role"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "content",
                    "role"
                ],
                "title": "ChatCompletionUserMessageParam"
            },
...
            "EmbeddingRequest": {
                "properties": {
                    "input": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "integer"
                                },
                                "type": "array"
                            },
                            {
                                "items": {
                                    "items": {
                                        "type": "integer"
                                    },
                                    "type": "array"
                                },
                                "type": "array"
                            },
                            {
                                "type": "string"
                            },
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            }
                        ],
                        "title": "Input"
                    },
                    "encoding_format": {
                        "anyOf": [
                            {
                                "type": "string",
                                "pattern": "^(float|base64)$"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Encoding Format",
                        "default": "float"
                    },
                    "model": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model"
                    },
                    "dimensions": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Dimensions"
                    },
                    "user": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "input"
                ],
                "title": "EmbeddingRequest"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
        },
        "securitySchemes": {
            "APIKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "api-key"
            }
        }
    }
}